x86/sgx: Introduce RECLAIM_IN_PROGRESS state
When a page is being reclaimed from the page pool (sgx_global_lru),
there is an intermediate stage where a page may have been identified
as a candidate for reclaiming, but has not yet been reclaimed.
Currently such pages are deleted from the global LRU, and stored in a
an array on stack. To prevent another thread from dropping the same
page in the middle of reclaiming, sgx_drop_epc_page() checks for
list_empty(page).

In future patches these pages need be stored in a temporary list that is
shared with multple cgroup reclaimers. so list_empty() should no longer be
used for this purpose. Add a RECLAIM_IN_PROGRESS state to explicitly
designate such intermediate stage of the reclaiming process. Do not drop
any page in this state in sgx_drop_epc_page().

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>
Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>
Cc: Sean Christopherson <seanjc@google.com>

V3:
- Extend the sgx_epc_page_state enum introduced earlier to replace the
flag based approach.
- Revised the commit message.