x86/sgx: Add 'struct sgx_epc_lru_lists' to encapsulate lru list(s)
Introduce a data structure to wrap the existing reclaimable list
and its spinlock in a struct to minimize the code changes needed
to handle multiple LRUs as well as reclaimable and non-reclaimable
lists. The new structure and helpers are not utilized in this
patch, but will used in a following set of patches to implement
SGX EPC cgroups.

This mirrors the implementation of parallel LRU lists for cgroups
in the core VM.

The changes to the structure needed for unreclaimable lists are
not included in this patch and will be added in a patch where they
are utilized.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>
Cc: Sean Christopherson <seanjc@google.com>