x86/sgx: Store reclaimable epc pages in sgx_epc_lru_lists
When an OOM event occurs, all pages associated with an enclave will
need to be freed, including pages that are not currently tracked by
the reclaimer. This means that each page will eventually need to be
added to the cgroup's LRU list struct whether it is tracked by the
reclaimer or not.

This patch prepares for adding the currently untracked pages by
replacing the Tracking pages that are not tracked by the reclaimer
by replacing functions sgx_mark_page_reclaimable() and
sgx_unmark_page_reclaimable() with sgx_record_epc_page() and
sgx_drop_epc_page(). sgx_record_epc_page() wil add the epc_page
to the "reclaimable" list in the sgx_epc_lru_lists struct.
sgx_drop_epc_page() will delete the page from the LRU list.

Currently this is a straight forward replacement of the two functions
for pages tracked by the reclaimer, and a follow on patch will add
the functionality to track unreclaimable pages using the same
functions.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>
Cc: Sean Christopherson <seanjc@google.com>