x86/sgx: Add EPC page flags to identify type of page
Unreclaimable pages can be either Version Array (VA) pages, or SECS
pages, which are Enclave pages.  Reclaimable pages must be Enclave pages.

Each sgx_epc_page has a value for "owner", which is stored in the
encl_owner field at allocation time. This value is a union, and its
meaning can change depending on what kind of page was allocated.
VA pages will have the owner field set to the owning enclave. Enclave
pages will have the owner set to the sgx_encl_page struct.

When cgroup OOM support is added in a later patch, the owning enclave
of a page will need to be identified from the list of unreclaimable
pages in the sgx_epc_lru_lists struct associated with the cgroup that
is out of memory.  Retrieving the sgx_encl struct from the page will
be different if the page is a VA page vs. an Enclave page.

Add 2 flags which will identify whether a page is an enclave page, or a
version array page so that the type of the encl_owner union can be
determined, and use them when adding newly allocated pages to the
unreclaimable or reclaimable lists during sgx_record_epc_page().

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>
Cc: Sean Christopherson <seanjc@google.com>