x86/sgx: Add EPC page flags to identify owner type
Each sgx_epc_page has a value for "owner", which is stored in the owner
field at allocation time. This value is a union, and its meaning can
change depending on what kind of page was allocated.  VA pages will have
the owner field set to the owning enclave. Enclave pages will have the
owner set to the sgx_encl_page struct.

When cgroup OOM support is added in a later patch, the owning enclave
of a page will need to be identified. Retrieving the sgx_encl struct
from a sgx_epc_page will be different if the page is a VA page vs. other
enclave pages.

Add 2 flags which will identify the type of the owner and apply them
accordingly to newly allocated pages.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>
Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>
Cc: Sean Christopherson <seanjc@google.com>

V3:
- Renamed the flags to clarify they are used to identify the struct type
of the owner.
- Revised commit message.