x86/cet: Handle #CP inside enclave
An enclave with shadow stacks enabled can trigger a #CP within the
enclave, just like a normal application with shadow stacks enabled. An
enclave triggered exception can only be handled by the enclave itself.
Properly route enclave triggered #CPs and enable enclave access to the
additional #CP info needed for handling.

Previously, a special exception fixup interface was designed and
implemented in the vDSO to route all enclave exceptions (and only
enclave exceptions) to user space instead of using the regular signal
mechanism. Modify the CET fault handler to follow this existing design
and use the vDSO fixup interface to route enclave triggered #CPs to user
space so the enclave can handle its own #CPs as needed. Note the fixup
only handles the #CPs if it determines that the exceptions are triggered
by an enclave so it does not interfere with handling of non-enclave
exceptions.

An enclave wishes to handle its own #CP also needs to opt-in by
setting the SECS.MISSELECT.CPINFO bit so that #CP info is reported in
the state saving area (SSA) that the enclave can access and use to
handle the exception. Define this bit and remove it from reserved mask
if the kernel sets CR4.CET bit.

Co-developed-by: Junjun Gu <junjun.gu@intel.com>
Signed-off-by: Junjun Gu <junjun.gu@intel.com>
Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>