x86/sgx: Prepare for EAUGing shadow stack pages
Modify sgx_encl_eaug_page function to handle cases for shadow stack page
types.

Move the allocation of encl_page object out of this function to the
caller so that the caller can set types of the EPC pages to be EAUG'd.

Additionally, failed attempt on EAUG can succeed in retries on next #PF.
So logic for inserting the allocated encl_page object into the enclave
bookkeeping xarray is also moved out and the encl_page for EAUG is kept
alive once insertion succeeds regardless of EAUG results.

Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>