x86/sgx: Enable SGX CET attributes
This patch is to pepare for enabling CET features
inside enclave. It removes the CET bit(6) in SECS.attributes
from reserved bitmask so it can be used by enclaves.

For an enclave to use CET features, it must enable CET
attributes by setting that bit. Also for each individual CET
feature an enclave wishes to opt in, it needs to set the
corresponding bits in the SECS.cet_attributes field, e.g., bit
0 for shadow stack, bit 1 for WRSS, bit 2 for IBT, etc.

Related fields in SECS, TCS, SIGSTRUCT are also added.

Co-developed-by: Junjun Gu <junjun.gu@intel.com>
Signed-off-by: Junjun Gu <junjun.gu@intel.com>
Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>