x86/sgx: Enable shadow stack in enclaves
To use shadow stack, an enclave needs to config and add
two types of EPC pages: PT_SS_FIRST to be used as the first
page of a shadow stack, PT_SS_REST for other pages
in a shadow stack.

On an async exit (AEX) event, CET states including the
current SSP are saved in an area configured as CET_SSA in
the TCS during EENTER. Page type for CET_SSA must be
PT_SS_REST and is only accessible by CPU during EENTER/ERESUME/AEX.
Such an access is not considered as a shadow stack access instead
a normal read or write by CPU, and the PFEC.shstk bit won't be set
in case of #PF triggered.

For normal shadow stack pages (PT_SS_FIRST and PT_SS_REST not used
for CET_SSA), the SGX driver needs to configure the VMAs with
VM_SHADOW_STACK flag just like regular shadow stack pages used outside
enclaves. For CET_SSA pages the driver needs to configure it as PT_SS_REST
in EPCM but mark the VMA as regular page with VM_READ|VM_WRITE.

To enable user space to mark a page to be CET_SSA, a new page control
flag, SGX_PAGE_CET_SSA is added and can be passed in the 'flags' field
of the EADD ioctl() parameters. The driver then records this info
by setting bit 8 (never to be used for real page types)  of the
encl_page->type. Later during mmap for these pages, the VM_SHADOW_STACK
flag is set according to bits in encl_page->type.

Co-developed-by: Junjun Gu <junjun.gu@intel.com>
Signed-off-by: Junjun Gu <junjun.gu@intel.com>
Signed-off-by: Haitao Huang <haitao.huang@linux.intel.com>